backend/requirements.txt
Flask
spotipy
fpdf
flask-cors
python-dotenv